Reorganize walk-in feature filesystem
