+++ apps/web/src/features/walk-in/onboarding/config.ts
