+++ apps/web/src/features/walk-in/onboarding/models/lock-in.ts
@@ -1,4 +1,4 @@
-import { formatDurationMinutes } from "./walk-in-onboarding-format";
+import { formatDurationMinutes } from "../format";
 
 type LockInStageState =
 	| "missing"
