+++ apps/web/src/features/walk-in/onboarding/models/quality.ts
@@ -1,4 +1,4 @@
-import { clampNumber, formatPercent } from "./walk-in-onboarding-format";
+import { clampNumber, formatPercent } from "../format";
 
 type QualityStageState =
 	| "missing"
