+++ apps/web/src/features/walk-in/onboarding/stages/metrics.tsx
@@ -12,8 +12,8 @@ import {
 	resolveRepoPulseStageModel,
 	resolveScalePreviewTokens,
 	resolveScaleStageModel,
-} from "./walk-in-onboarding-models";
-import type { WalkInOnboardingMetrics } from "./walk-in-onboarding-types";
+} from "../models";
+import type { WalkInOnboardingMetrics } from "../types";
 
 interface SharedStageProps {
 	onboardingMetrics: WalkInOnboardingMetrics;
