+++ apps/web/src/features/walk-in/onboarding/stages/model.tsx
@@ -4,8 +4,8 @@ import {
 	getModelStageTone,
 	resolveModelPreviewInput,
 	resolveModelStageModel,
-} from "./walk-in-onboarding-models";
-import type { WalkInOnboardingMetrics } from "./walk-in-onboarding-types";
+} from "../models";
+import type { WalkInOnboardingMetrics } from "../types";
 
 interface ModelStageProps {
 	onboardingMetrics: WalkInOnboardingMetrics;
