+++ apps/web/src/features/walk-in/onboarding/stages/skills.tsx
@@ -7,8 +7,8 @@ import {
 	getSkillsCardStyle,
 	resolveSkillsPreviewInput,
 	resolveSkillsStageModel,
-} from "./walk-in-onboarding-models";
-import type { WalkInOnboardingMetrics } from "./walk-in-onboarding-types";
+} from "../models";
+import type { WalkInOnboardingMetrics } from "../types";
 
 interface SkillsStageProps {
 	onboardingMetrics: WalkInOnboardingMetrics;
