+++ apps/web/src/features/walk-in/team-card/row.ts
